# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API used for manages school environment.",
        "title": "School Information System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/permissions": {
            "get": {
                "description": "Admin with permission read permission resource only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permission"
                ],
                "summary": "Get existing permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "Cookie",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "refresh_token",
                        "name": "Cookie2",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "create",
                                "read",
                                "update",
                                "delete"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "role",
                                "permission"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "resource",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/swaglib.Envelope"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "data": {
                                                "$ref": "#/definitions/models.Permission"
                                            },
                                            "meta": {
                                                "$ref": "#/definitions/swaglib.Pagination"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/swaglib.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Admin with permission create permission resource only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permission"
                ],
                "summary": "Create new permission combination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "Cookie",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "refresh_token",
                        "name": "Cookie2",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "data of new permission",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payloads.RequestCreatePermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/swaglib.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Permission"
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/swaglib.Info"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/swaglib.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/permissions/grant": {
            "put": {
                "description": "Admin with permission update permission resource only. Response granted user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permission"
                ],
                "summary": "Grant existing permission to another admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "Cookie",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "refresh_token",
                        "name": "Cookie2",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "data to grant permission",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payloads.RequestGrantPermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/swaglib.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/models.User"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "admin_profile": {
                                                            "allOf": [
                                                                {
                                                                    "$ref": "#/definitions/models.Admin"
                                                                },
                                                                {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "permissions": {
                                                                            "type": "array",
                                                                            "items": {
                                                                                "$ref": "#/definitions/models.Permission"
                                                                            }
                                                                        }
                                                                    }
                                                                }
                                                            ]
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/swaglib.Info"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/swaglib.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/permissions/revoke": {
            "put": {
                "description": "Admin with permission update permission resource only. Response granted user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "permission"
                ],
                "summary": "Revoke existing permission of another admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "Cookie",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "refresh_token",
                        "name": "Cookie2",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "data to revoke permission",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payloads.RequestRevokePermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/swaglib.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/models.User"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "admin_profile": {
                                                            "allOf": [
                                                                {
                                                                    "$ref": "#/definitions/models.Admin"
                                                                },
                                                                {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "permissions": {
                                                                            "type": "array",
                                                                            "items": {
                                                                                "$ref": "#/definitions/models.Permission"
                                                                            }
                                                                        }
                                                                    }
                                                                }
                                                            ]
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/swaglib.Info"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/swaglib.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ErrorPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Admin": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "joined_at": {
                    "type": "string"
                },
                "staff_role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Permission": {
            "type": "object",
            "properties": {
                "actions": {
                    "description": "[](\"create\", \"read\", \"update\", \"delete\")",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PermissionAction"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "resource": {
                    "$ref": "#/definitions/models.PermissionResource"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.PermissionAction": {
            "type": "string",
            "enum": [
                "create",
                "read",
                "update",
                "delete"
            ],
            "x-enum-varnames": [
                "ActionCreate",
                "ActionRead",
                "ActionUpdate",
                "ActionDelete"
            ]
        },
        "models.PermissionResource": {
            "type": "string",
            "enum": [
                "role",
                "permission"
            ],
            "x-enum-varnames": [
                "ResourceRole",
                "ResourcePermission"
            ]
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "archived_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "auth username",
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "description": "\"male\", \"female\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserGender"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "phone": {
                    "description": "phone number",
                    "type": "string"
                },
                "role": {
                    "description": "\"student\", \"teacher\", \"admin\", \"unsetted\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserRole"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.UserGender": {
            "type": "string",
            "enum": [
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale"
            ]
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "student",
                "teacher",
                "admin",
                "unsetted"
            ],
            "x-enum-varnames": [
                "RoleStudent",
                "RoleTeacher",
                "RoleAdmin",
                "RoleUnsetted"
            ]
        },
        "payloads.RequestCreatePermission": {
            "type": "object",
            "required": [
                "actions",
                "description",
                "name",
                "resource"
            ],
            "properties": {
                "actions": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.PermissionAction"
                    }
                },
                "description": {
                    "type": "string",
                    "minLength": 10
                },
                "name": {
                    "type": "string",
                    "minLength": 10
                },
                "resource": {
                    "enum": [
                        "role",
                        "permission"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PermissionResource"
                        }
                    ]
                }
            }
        },
        "payloads.RequestGrantPermission": {
            "type": "object",
            "required": [
                "permission_id",
                "target_id"
            ],
            "properties": {
                "permission_id": {
                    "type": "string"
                },
                "target_id": {
                    "type": "string"
                }
            }
        },
...
        "payloads.RequestRevokePermission": {
            "type": "object",
            "required": [
                "permission_id",
                "target_id"
            ],
            "properties": {
                "permission_id": {
                    "type": "string"
                },
                "target_id": {
                    "type": "string"
                }
            }
        },
...
        "reply.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "details": {
                    "description": "Optional debug details",
                    "type": "string"
                },
                "fields": {
                    "description": "Fields causing the error (if any)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                }
            }
        },
...
        "swaglib.Envelope": {
            "type": "object",
            "properties": {
                "data": {
                    "x-null-if-{}": true
                },
                "meta": {
                    "$ref": "#/definitions/swaglib.Meta"
                }
            }
        },
        "swaglib.Info": {
            "type": "object",
            "properties": {
                "information": {
                    "description": "information message",
                    "type": "string",
                    "example": "resource successfully updated"
                }
            }
        },
        "swaglib.Meta": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/swaglib.status"
                }
            }
        },
        "swaglib.Pagination": {
            "type": "object",
            "properties": {
                "pagination": {
                    "type": "object",
                    "properties": {
                        "current": {
                            "description": "current offset",
                            "type": "integer"
                        },
                        "hasNext": {
                            "description": "true if data more than replied",
                            "type": "boolean"
                        },
                        "next": {
                            "description": "if hasNext is false, next is 0",
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "swaglib.status": {
            "type": "string",
            "enum": [
                "SUCCESS",
                "ERROR"
            ],
            "x-enum-varnames": [
                "success",
                "error"
            ]
        }
    }
}